# docs/swagger.yaml
basePath: /
definitions:
  model.DeletePlayerErrorResponse:
    properties:
      deleted:
        type: boolean
      error:
        type: string
      pid:
        type: string
    type: object
  model.DeletePlayerResponse:
    properties:
      deleted:
        type: boolean
      message:
        type: string
      pid:
        type: string
    type: object
  model.PlayerAPIResponse:
    properties:
      message:
        type: string
      player:
        $ref: '#/definitions/model.PlayerInfo'
      type:
        type: integer
    type: object
  model.PlayerInfo:
    properties:
      alliance:
        type: string
      dName:
        type: string
      kid:
        type: integer
      pfp:
        type: string
      pid:
        type: integer
    type: object
  model.PlayerReqBody:
    properties:
      alliance:
        type: string
      kid:
        type: integer
      pid:
        type: string
    required:
    - pid
    type: object
  model.UpdatePlayerSSE:
    properties:
      player:
        $ref: '#/definitions/model.PlayerInfo'
      updated:
        type: boolean
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for redeeming KingShot gift codes on behalf of registered players.
    OpenAPI spec served at /api/openapi.json for import into GitBook.
  title: Kingshot Redeem API
  version: "1.0"
paths:
  /:
    get:
      description: Returns "Hello, World!" to confirm the service is running.
      produces:
      - text/plain
      responses:
        "200":
          description: Hello, World!
          schema:
            type: string
      summary: Health check
      tags:
      - System
  /player/add:
    post:
      consumes:
      - application/json
      description: Adds a new player to the database. When KSFetchAvailable is enabled,
        profile data is fetched from the KingShot API; otherwise only the provided
        fields are stored.
      parameters:
      - description: Player to add
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/model.PlayerReqBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created Player examples" examples(
          schema:
            $ref: '#/definitions/model.PlayerAPIResponse'
        "400":
          description: Invalid JSON or validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Unable to add the player
          schema:
            additionalProperties: true
            type: object
      summary: Add a player
      tags:
      - Player
  /player/del/{pid}:
    delete:
      description: Removes a player from the database by player ID.
      parameters:
      - description: Player ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Player deleted" example({"pid":123,"deleted":true})
          schema:
            $ref: '#/definitions/model.DeletePlayerResponse'
        "500":
          description: Delete failed" example({"pid":123,"deleted":false,"error":"database
            error"})
          schema:
            $ref: '#/definitions/model.DeletePlayerErrorResponse'
      summary: Delete a player
      tags:
      - Player
  /player/update/:
    patch:
      description: Updates a single Player with the playerID provided in cia teh param
      parameters:
      - description: Player ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdatePlayerSSE'
      summary: Updates the Db with latest Player Data of that player with same PlayerID
      tags:
      - Player
  /redeem/db:
    post:
      description: 'Loads all players from the database and redeems the configured
        gift code for each one, streaming per-player results as Server-Sent Events
        (event: redeem-res) followed by a summary (event: redeem-fin). Rate limited.'
      parameters:
      - description: Gift code
        example: KS0803
        in: query
        name: code
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'redeem-res and redeem-fin event" example(event: redeem-res\ndata:
            {"fid":"123","code":"KS0803","result":{"errCode":20000,"msg":"Redeemed"},"success":true,"time":"2026-08-07T22:05:40+05:30"}\n\n
            event: redeem-fin\ndata: {"redeemed":10,"manual":2,"playerDead":1,"codeExpired":3})'
          schema:
            type: string
      summary: Redeem gift code for all players
      tags:
      - Redeem
  /redeem/ws:
    get:
      description: Upgrades the connection to a WebSocket and greets the client.
      responses:
        "101":
          description: Switching Protocols
      summary: WebSocket test
      tags:
      - Redeem
swagger: "2.0"